added validation to new problems
problems now require a title, and uploads must be images. I'd like to
change it so that uploads can be images or movies, but I'm not sure how
to do that. Also, I want to setup the validator to require at least one
tag, but need to work on that.

Finally, I need to change the behavior of the validator. Right now it
just redirects to /, but I want it to mark the problem fields with
explanation text.